Refactor SAP Systems code
